requirements.txt
streamlit==1.37.0
python-dotenv==1.0.1
openai==1.40.0
transformers==4.43.3
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0
sentencepiece
pypdf2==3.0.1
python-docx==1.1.2
